requirements.txt
# Django Core
Django==5.2.1
django-cors-headers==4.3.1

# Django REST Framework
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Filters and Search
django-filter==24.2

# Database (for production PostgreSQL support)
psycopg2-binary==2.9.9

# Environment Variables
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.2

# Testing
pytest==8.2.0
pytest-django==4.8.0
factory-boy==3.3.0

# Code Quality
flake8==7.0.0
black==24.4.2

# Security
django-ratelimit==4.1.0

# Production Server
gunicorn==22.0.0
whitenoise==6.6.0
